requirements.txt
python-dotenv
flask
Flask-Cors
blueprint
numpy
mysql-connector-python
pymysql
pandas
openpyxl
xlsxwriter
pdfkit
lxml
matplotlib